Move the search history option logics
