+++ translations/en.json
@@ -58,5 +58,6 @@
   "saveName": "Enter a name",
   "saveDescription": "Enter a description",
   "clearHistoryMsg": "Are you sure you want to clear the search history?",
-  "deleteHistoryItem": "Are you sure you want to delete {{searchItemName}} from the search history?`"
+  "deleteHistoryItem": "Are you sure you want to delete {{searchItemName}} from the search history?",
+  "searchIsCompleted": "Search is completed with {{numOfResults}} result(s)"
 }
